It's been a while

+++ typewriter/typewriter.sh
@@ -1,14 +1,26 @@
 #!/bin/bash
 
-echo "Welcome to typewriter.sh, a SMUT shim-building automation tool"
-echo "Made by rainestorme"
+if [ "$EUID" -ne 0 ]; then 
+  echo "Please run as root"
+  exit
+fi
 
 echo "Before you start, confirm that you have the following files in the /typewriter folder:"
 echo " - raw_shim.bin - your raw shim for the board you want to build for"
 echo " - reco_image.bin - your v105 recovery image for the board you want to build for"
-echo "Also make sure this script was run as root! (TODO add root check)"
-echo " Press enter to continue, Ctrl+C to exit"
-read -r
+
+if [ ! -f ./reco_image.bin ];
+  echo "Please place reco_image.bin in the current directory."
+  exit 1
+fi
+
+if [ ! -f ./raw_shim.bin ];
+  echo "Please place raw_shim.bin in the current directory."
+  exit 1
+fi
+
+echo "Welcome to typewriter.sh, a SMUT shim-building automation tool"
+echo "Made by rainestorme"
 
 echo "Copying reco image (this will take a while)..."
 cp ./reco_image.bin ./murkmod_image.bin
